Move sys tests to sys_tx_tests.rs
