+++ module-system/module-implementations/sov-evm/src/tests/mod.rs
@@ -3,6 +3,7 @@ mod cfg_tests;
 mod genesis_tests;
 mod hooks_tests;
 mod queries;
+mod sys_tx_tests;
 pub(crate) mod test_signer;
 mod tx_tests;
